requirements.txt
openpyxl>=3.1.5
pandas>=2.2.3
playwright>=1.49.0
requests>=2.32.3
numpy>=2.1.3
linkedin-api>=2.3.1
pymysql
minio
xlrd
s3fs
pyiceberg[s3fs,hive,glue]
python-dateutil